removed unnecessary columns (stil not the feature data) and checked for missing values.
removed unnecessary columns, the data set is still not the feuture data set with only numeric values since it can we usefull to have certain columns to filter on or check the country or subid for example. Also checked for missing values. lost of missing values in the chemicals properties and emission data. this is because not al this data can be joined since not al the CAS# in the various datasets overlap.  At the moment i am not sure what to do with this data since removing al the missings will result in a huge data loss , but imputing this data would not make a lot of sence I think.